+++ ui-test/tests/Icons.jsx
@@ -42,6 +42,11 @@ const TestIcons = React.createClass({
                         name="non_existing_icon_name"
                         tooltip="Icon with non existent icon name"
                     />
+                    <Icon
+                        name="account_box"
+                        badge="5"
+                        tooltip="Icon with badge"
+                    />
                 </CardContent>
             </Card>
         );